+++ src/models/index.ts
@@ -1,1 +1,2 @@
+export * from './ContentData';
 export * from './Theme';