+++ src/webview/components/EmptyPlaceholder.tsx
@@ -1,0 +1,13 @@
+import * as React from 'react';
+
+export interface IEmptyPlaceholderProps {}
+
+export const EmptyPlaceholder: React.FunctionComponent<IEmptyPlaceholderProps> = (props: React.PropsWithChildren<IEmptyPlaceholderProps>) => {
+  return (
+    <div className='mt-24 text-xl flex justify-center flex-col space-y-12'>
+      <p>⬅</p>
+      <p>⬅ Please select some Markdown content first</p>
+      <p>⬅</p>
+    </div>
+  );
+};